backend/modules/plagiarism/writingstyleanalaysis/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
scikit-learn
numpy
joblib
mysql-connector-python
requests
ddgs
trafilatura
playwright
urllib3
